+++ src/main/java/com/ssginc8/docto/patient/controller/PatientController.java
@@ -1,11 +1,10 @@
 package com.ssginc8.docto.patient.controller;
 
-import com.ssginc8.docto.patient.dto.PatientRequestDTO;
-import com.ssginc8.docto.patient.dto.PatientResponseDTO;
+import com.ssginc8.docto.patient.dto.PatientRequest;
+import com.ssginc8.docto.patient.dto.PatientResponse;
 import com.ssginc8.docto.patient.service.PatientService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,12 +17,12 @@ public class PatientController {
 	private final PatientService patientService;
 
 	@PostMapping
-	public ResponseEntity<PatientResponseDTO> create(@RequestBody PatientRequestDTO dto) {
+	public ResponseEntity<PatientResponse> create(@RequestBody PatientRequest dto) {
 		return ResponseEntity.ok(patientService.createPatient(dto));
 	}
 
 	@GetMapping
-	public ResponseEntity<List<PatientResponseDTO>> getAll() {
+	public ResponseEntity<List<PatientResponse>> getAll() {
 		return ResponseEntity.ok(patientService.getAllPatients());
 	}
