+++ src/main/java/com/ssginc8/docto/patient/repo/PatientRepo.java
@@ -1,13 +1,12 @@
-package com.ssginc8.docto.patient.repository;
+package com.ssginc8.docto.patient.repo;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssginc8.docto.patient.entity.Patient;
 
-public interface PatientRepository extends JpaRepository<Patient, Long> {
+public interface PatientRepo extends JpaRepository<Patient, Long> {
 	Optional<Patient> findByPatientIdAndDeletedAtIsNull(Long patientId);
 
 }
